requirements.txt
google-adk
litellm[proxy]
litellm
python-dotenv
openai
certifi
